Revert "added fuctionality"
