+++ src/components/navbar/Navebar.tsx
@@ -51,10 +51,6 @@ function TopNavebar() {
     }
   }
 
-  const handleClick = () => {
-    navigate("/chat")
-  }
-
   return (
     <div>
       <div style={{ paddingBottom: '25px' }}>
@@ -109,10 +105,7 @@ function TopNavebar() {
               title={user["displayName"]}
               menuVariant="light"
             >
-             
-
               {!isAdmin && <NavDropdown.Item>Messages</NavDropdown.Item>}
-
               <NavDropdown.Item onClick={Logout}>Log Out</NavDropdown.Item>
               
             </NavDropdown>
