--- src/pages/chatContext.d.ts
@@ -1,5 +1,0 @@
-declare module "./chatContext"{
-    export const getChat: any;
-}
-
-export{}
